2010-11-20  Tatsuhiro Tsujikawa  <t-tujikawa@users.sourceforge.net>
	Use format specifier %lld for cuid_t. Defined cuid_t as long long
	int instead of int64_t, since g++ complains int64_t is not
	suitable for %lld.

+++ src/Command.h
@@ -39,7 +39,7 @@
 
 namespace aria2 {
 
-typedef int64_t cuid_t;
+typedef long long int cuid_t;
 
 class Command {
 public:
